+++ modules/drivers/velodyne/driver/input.cc
@@ -47,18 +47,22 @@ bool Input::exract_nmea_time_from_packet(NMEATimePtr nmea_time,
     }
   }
 
-  nmea_time->year = (bytes[date_field_index + 4] - '0') * 10 +
-                    (bytes[date_field_index + 5] - '0');
-  nmea_time->mon = (bytes[date_field_index + 2] - '0') * 10 +
-                   (bytes[date_field_index + 3] - '0');
-  nmea_time->day = (bytes[date_field_index] - '0') * 10 +
-                   (bytes[date_field_index + 1] - '0');
-  nmea_time->hour = (bytes[time_field_index] - '0') * 10 +
-                    (bytes[time_field_index + 1] - '0');
-  nmea_time->min = (bytes[time_field_index + 2] - '0') * 10 +
-                   (bytes[time_field_index + 3] - '0');
-  nmea_time->sec = (bytes[time_field_index + 4] - '0') * 10 +
-                   (bytes[time_field_index + 5] - '0');
+  nmea_time->year =
+      static_cast<uint16_t>((bytes[date_field_index + 4] - '0') * 10 +
+                            (bytes[date_field_index + 5] - '0'));
+  nmea_time->mon =
+      static_cast<uint16_t>((bytes[date_field_index + 2] - '0') * 10 +
+                            (bytes[date_field_index + 3] - '0'));
+  nmea_time->day = static_cast<uint16_t>((bytes[date_field_index] - '0') * 10 +
+                                         (bytes[date_field_index + 1] - '0'));
+  nmea_time->hour = static_cast<uint16_t>((bytes[time_field_index] - '0') * 10 +
+                                          (bytes[time_field_index + 1] - '0'));
+  nmea_time->min =
+      static_cast<uint16_t>((bytes[time_field_index + 2] - '0') * 10 +
+                            (bytes[time_field_index + 3] - '0'));
+  nmea_time->sec =
+      static_cast<uint16_t>((bytes[time_field_index + 4] - '0') * 10 +
+                            (bytes[time_field_index + 5] - '0'));
 
   if (nmea_time->year < 0 || nmea_time->year > 99 || nmea_time->mon > 12 ||
       nmea_time->mon < 1 || nmea_time->day > 31 || nmea_time->day < 1 ||
