+++ modules/drivers/velodyne/parser/velodyne_parser.cc
@@ -24,13 +24,13 @@ namespace drivers {
 namespace velodyne {
 
 uint64_t VelodyneParser::GetGpsStamp(double current_packet_stamp,
-                                       double *previous_packet_stamp,
-                                       uint64_t *gps_base_usec) {
+                                     double *previous_packet_stamp,
+                                     uint64_t *gps_base_usec) {
   if (current_packet_stamp < *previous_packet_stamp) {
     // plus 3600 when large jump back, discard little jump back for wrong time
     // in lidar
     if (std::abs(*previous_packet_stamp - current_packet_stamp) > 3599000000) {
-      *gps_base_usec += 3600 * 1e6;
+      *gps_base_usec += static_cast<uint64_t>(3600 * 1e6);
       AINFO << "Base time plus 3600s. Model: " << config_.model() << std::fixed
             << ". current:" << current_packet_stamp
             << ", last time:" << *previous_packet_stamp;
@@ -47,7 +47,8 @@ uint64_t VelodyneParser::GetGpsStamp(double current_packet_stamp,
   }
 
   *previous_packet_stamp = current_packet_stamp;
-  uint64_t gps_stamp = *gps_base_usec + current_packet_stamp;
+  uint64_t gps_stamp =
+      *gps_base_usec + static_cast<uint64_t>(current_packet_stamp);
 
   gps_stamp = gps_stamp * 1000;
   return gps_stamp;
@@ -63,7 +64,7 @@ PointXYZIT VelodyneParser::get_nan_point(uint64_t timestamp) {
   return nan_point;
 }
 
-VelodyneParser::VelodyneParser(const Config& config)
+VelodyneParser::VelodyneParser(const Config &config)
     : last_time_stamp_(0), config_(config), mode_(STRONGEST) {}
 
 void VelodyneParser::init_angle_params(double view_direction,
@@ -123,9 +124,8 @@ bool VelodyneParser::is_scan_valid(int rotation, float range) {
 }
 
 void VelodyneParser::ComputeCoords(const float &raw_distance,
-                                    const LaserCorrection &corrections,
-                                    const uint16_t rotation,
-                                    PointXYZIT *point) {
+                                   const LaserCorrection &corrections,
+                                   const uint16_t rotation, PointXYZIT *point) {
   // ROS_ASSERT_MSG(rotation < 36000, "rotation must between 0 and 35999");
   assert(rotation <= 36000);
   double x = 0.0;
